requirements.txt
groq>=0.18.0
requests>=2.31.0
python-dotenv>=1.0.0
typing-extensions>=4.5.0
flask>=2.0.1
geopy>=2.3.0
overpy>=0.6
folium>=0.12.1
geocoder>=1.38.1
